requirements.txt
google-adk
fpdf
markdown2